1. Move tool's type definition to ToolController_ folder 2. Update code to use getToolsForPanel function
